Improve load speed
glm_init
mm_intit
fem_init
