+++ Project.toml
@@ -1,7 +1,7 @@
 name = "MixedAnova"
 uuid = "72e9ff3b-fbb7-4ec3-b65d-9c3416869820"
 authors = ["Jejulia <Jejulia@users.noreply.github.com> and contributors"]
-version = "0.2.2"
+version = "0.3.0"
 
 [deps]
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
