+++ src/GLM/anova.jl
@@ -1,6 +1,34 @@
 # ===========================================================================================
 # Main API
 
+using GLM
+@reexport using GLM
+import GLM: glm, LinPredModel, LinearModel, LmResp, DensePred, DensePredChol, SparsePredChol, QRCompactWY, LinPred, installbeta!, delbeta!,  linpred!,
+            updateμ!, linpred, cholfactors, updateμ!,  AbstractGLM, FP, SparseMatrixCSC, Link
+
+const FixDispDist = Union{Bernoulli, Binomial, Poisson}
+
+"""
+    canonicalgoodnessoffit(::FixDispDist) = LRT
+    canonicalgoodnessoffit(::UnivariateDistribution) = FTest
+
+    const FixDispDist = Union{Bernoulli, Binomial, Poisson}
+    
+Return LRT if the distribution has fixed dispersion
+"""
+canonicalgoodnessoffit(::FixDispDist) = LRT
+canonicalgoodnessoffit(::UnivariateDistribution) = FTest
+
+"""
+    glm(f, df::DataFrame, d::Binomial, l::GLM.Link, args...; kwargs...)
+
+Automatically transform dependent variable into 0/1 for family `Binomial`
+"""
+glm(f::FormulaTerm, df::DataFrame, d::Binomial, l::Link, args...; kwargs...) = 
+    fit(GeneralizedLinearModel, f, 
+        combine(df, : , f.lhs.sym => ByRow(x -> x == unique(df[:, f.lhs.sym])[end]) => f.lhs.sym), 
+        d, l, args...; kwargs...)
+
 """
     anova(<models>...; test::Type{T}) where {T <: GoodnessOfFit}
 
@@ -25,11 +53,6 @@ anova(models::Vararg{TableRegressionModel{<: LinearModel, <: AbstractArray}, N};
         kwargs...) where {N, T <: GoodnessOfFit} = 
     anova(test, models...; kwargs...)
 
-anova(models::Vararg{LinearMixedModel, N}; 
-        test::Type{T} = length(models) > 1 ? (LRT) : (FTest), 
-        kwargs...) where {N, T <: GoodnessOfFit} = 
-    anova(test, models...; kwargs...)
-
 anova(models::Vararg{TableRegressionModel{<: GeneralizedLinearModel, <: AbstractArray}, N}; 
         test::Type{T} = canonicalgoodnessoffit(models[1].model.rr.d),
         kwargs...) where {N, T <: GoodnessOfFit} = 
@@ -37,6 +60,7 @@ anova(models::Vararg{TableRegressionModel{<: GeneralizedLinearModel, <: Abstract
 
 # ==================================================================================================================
 # ANOVA by F test 
+# LinearModels
 
 """
     anova(::Type{FTest}, <model>; kwargs...)
@@ -64,57 +88,6 @@ function anova(::Type{FTest},
     fstat = (MSR[1:(end-1)] / last(MSR)..., NaN)
     pvalue = (ccdf.(FDist.(df, last(df)), abs.(fstat))[1:(end-1)]..., NaN)
     AnovaResult(model, FixedAnovaStatsF{LinearModel, length(df)}(type, nobs(model), tuple(df...), ss, fstat, pvalue))
-end
-
-# ----------------------------------------------------------------------------------------------------
-# Linear mixed-effect models
-
-function anova(::Type{FTest}, 
-            model::LinearMixedModel; 
-            type::Int = 1, 
-            adjust_sigma::Bool = true)
-
-    @assert (type in [1,2,3]) "Invalid type"
-    @assert (type in [1,3]) "Type 2 anova is not supported now"
-
-    varβ = vcov(model) 
-    β = fixef(model)
-
-    assign = asgn(first(model.formula.rhs))
-    intercept = first(assign) == 1
-
-    # calculate degree of freedom for factors and residuals
-    df, resdf, dofinfo = calcdof(model)
-    
-    # use MMatrix/SizedMatrix ?
-    if type == 1
-        invvarfixchol = cholesky(inv(varβ)|> Hermitian).L 
-        # adjust σ like linear regression
-        model.optsum.REML || adjust_sigma && begin
-            invvarfixchol = invvarfixchol / sqrt(nobs(model) / (nobs(model) - length(β)))
-        end 
-        fs = invvarfixchol'β
-        uniqas = unique(assign)
-        fstat = ntuple(lastindex(uniqas)) do fac
-            mapreduce(val->fs[val] ^ 2, +, findall(==(fac), assign)) / df[fac]
-        end
-    else 
-        # calculate block by block
-        adjust = 1.0
-        model.optsum.REML || adjust_sigma && (adjust = (nobs(model) - length(β)) / nobs(model)) 
-        offset = 0
-        intercept || (offset = 1)
-        fstat = ntuple(last(assign) - offset) do factor
-            select = findall(==(factor + offset), assign)
-            invvarfix = inv(varβ[select, select]) 
-            view(β, select)' * invvarfix * view(β, select) / rank(invvarfix) * adjust
-        end
-    end
-
-    pvalue = ntuple(lastindex(fstat)) do id
-            ccdf(FDist(df[id], resdf[id]), abs(fstat[id]))
-    end
-    AnovaResult(model, MixedAnovaStatsF{LinearMixedModel, length(fstat)}(type, nobs(model), df, resdf, fstat, pvalue, dofinfo))
 end
 
 # ----------------------------------------------------------------------------------------
@@ -150,6 +123,7 @@ end
 
 # ==================================================================================================================
 # ANOVA by Likehood-ratio test 
+# LinearModels
 
 """
     anova(::Type{LRT}, <model>; kwargs...)
@@ -181,25 +155,6 @@ function anova(::Type{LRT},
     AnovaResult(model, FixedAnovaStatsLRT{LinearModel, n}(1, nobs(model), df, tuple(dev...), lrstat, pval))
 end
 
-# ----------------------------------------------------------------------------------------------------------
-# ANOVA for LinearMixedModel
-
-function anova(::Type{LRT}, model::LinearMixedModel)
-    @warn "fit all submodels"
-    null = first(first(formula(model).rhs).terms) == InterceptTerm{false}()
-    models = nestedmodels(model; null = null)
-    anova(LRT, models)
-end
-
-function anova(::Type{LRT}, models::NTuple{N, LinearMixedModel}) where N
-    n = Int(nobs(first(models)))
-    df = dof.(models)
-    Δdf = _diff(df)
-    dev = deviance.(models)
-    lrstat = _diffn(dev)
-    pval = ccdf.(Chisq.(abs.(Δdf)), abs.(lrstat))
-    AnovaResult(models, MixedAnovaStatsLRT{LinearMixedModel, length(Δdf)}(1, n, Δdf, dev[2:end], lrstat, pval))
-end
 
 # ------------------------------------------------------------------------------------------------------------
 # ANOVA for GeneralizedLinearModel
@@ -247,15 +202,6 @@ function anova(::Type{FTest},
 end
 
 function anova(::Type{FTest}, 
-        models::Vararg{<: LinearMixedModel, N}; 
-        testnested::Bool = true) where N
-    
-    @warn "F test can only be appplied to a single linear mixed-effects model"
-    aov = anova(FTest, last(models))
-    AnovaResult(models, NestedAnovaStatsF{length(models)}(aov.stats.nobs, dof.(models), deviance.(models), (NaN, aov.stats.fstat...), (NaN, aov.stats.pval...)))
-end
-
-function anova(::Type{FTest}, 
         models::Vararg{TableRegressionModel{<: GeneralizedLinearModel, <: AbstractArray}, N}; 
         testnested::Bool = true) where N
 
@@ -283,20 +229,6 @@ function anova(::Type{LikelihoodRatioTest},
     Δdev = _diffn(dev)
     lrstat = (NaN, Δdev ./ σ² ...)
     pval = (NaN, ccdf.(Chisq.(Δdf), abs.(lrstat[2:end]))...)
-    AnovaResult(models, NestedAnovaStatsLRT{length(df)}(n, df, dev, lrstat, pval))
-end
-
-function anova(::Type{LikelihoodRatioTest}, 
-            models::Vararg{<: LinearMixedModel, N}; 
-            testnested::Bool = true) where N
-    # AIC and BIC
-    n = Int(nobs(first(models)))
-    df = dof.(models)
-    Δdf = _diff(df)
-    dev = deviance.(models)
-    Δdev = _diffn(dev)
-    lrstat = (NaN, Δdev...)
-    pval = (NaN, ccdf.(Chisq.(abs.(Δdf)), abs.(lrstat[2:end]))...)
     AnovaResult(models, NestedAnovaStatsLRT{length(df)}(n, df, dev, lrstat, pval))
 end
 
@@ -336,50 +268,6 @@ function anova(test::Type{T}, ::Type{LinearModel}, X, y;
         kwargs...) where {T <: GoodnessOfFit}
     model = lm(X, y, pivot; kwargs...)
     anova(test, model; pivot = pivot, type = type)
-end
-
-"""
-    anova_lme(f::FormulaTerm, tbl; test::Type{T} = FTest, <keyword arguments>)
-
-    anova_lme(test::Type{T}, f::FormulaTerm, tbl; <keyword arguments>)
-
-    anova(test::Type{T}, ::Type{LinearMixedModel}, f::FormulaTerm, tbl;
-            type::Int = 1, 
-            adjust_sigma::Bool = true)
-
-ANOVA for linear mixed-effect models.
-
-The arguments `f` and `tbl` are `Formula` and `DataFrame`.
-
-* `type` specifies type of anova. only `1, 3` are valid.
-* `adjust_sigma` determines whether adjust σ to match that of linear mixed-effect model fit by REML.
-
-Other keyword arguments
-* `wts = []`
-* `contrasts = Dict{Symbol,Any}()`
-* `verbose::Bool = false`
-* `REML::Bool = true`
-
-`anova_lme` generate a `LinearMixedModel` object through calling `anova`, which is fit by `lme` with REML.
-"""
-anova_lme(f::FormulaTerm, tbl; 
-        test::Type{T} = FTest,
-        kwargs...) where {T <: GoodnessOfFit} = 
-    anova(test, LinearMixedModel, f, tbl; kwargs...)
-
-anova_lme(test::Type{T}, f::FormulaTerm, tbl; 
-        kwargs...) where {T <: GoodnessOfFit} = 
-    anova(test, LinearMixedModel, f, tbl; kwargs...)
-
-function anova(test::Type{T}, ::Type{LinearMixedModel}, f::FormulaTerm, tbl; 
-        type::Int = 1, 
-        adjust_sigma::Bool = true,
-        wts = [], 
-        contrasts = Dict{Symbol,Any}(), 
-        verbose::Bool = false, 
-        REML::Bool = true) where {T <: GoodnessOfFit}
-    model = lme(f, tbl, wts = wts, contrasts = contrasts, verbose = verbose, REML = REML)
-    anova(test, model; type = type, adjust_sigma = adjust_sigma)
 end
 
 """
